pyproject.toml
[tool.poetry]
name = "pi-blur"
version = "0.1.0"
description = ""
authors = ["Chris Barry <chris@greatjones.co>"]

[[tool.poetry.source]]
name = "piwheels"
url = "https://www.piwheels.org/simple/"

[tool.poetry.scripts]
start = "pi_blur.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.25.1"
python-dotenv = "^0.15.0"
pillow = "^8.0.1"
inky = "^1.1.1"
"rpi.gpio" = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["setuptools", "wheel", "numpy"]
build-backend = "setuptools.build_meta:__legacy__"
